hw3/requirements.txt
pandas
numpy
tweet-preprocessor